% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACGCA-package.R
\docType{package}
\name{ACGCA-package}
\alias{ACGCA-package}
\title{ACGCA package for simulating tree growth and mortality}
\description{
The ACGCA package makes running the Allometrically Constrained Growth and 
Carbon Allocation model relatively easy. This package contains one function.
The function growthloopR() that runs the ACGCA model. Detailed documentation 
for the variables are available in the help file for the function and a 
demonstration of the functions use is available in the vignette accompanying
this package.
}
