% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACGCA.R
\docType{package}
\name{ACGCA}
\alias{ACGCA}
\alias{ACGCA-package}
\title{ACGCA: A package for simulating tree growth and mortality}
\description{
The ACGCA package makes running the Allometrically Constrained Growth and 
Carbon Allocation model relatively easy.
}
