% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACGCA_call_met.R
\name{ACGCArun}
\alias{ACGCArun}
\title{ACGCA modelfunction}
\usage{
ACGCArun(sparms, gparms, r0 = 0.05)
}
\arguments{
\item{sparms}{A vector containing the parameters for the simulationl.}

\item{gparms}{A vector containing values that control simulation behavior.}

\item{r0}{The starting radius. Defaults to 0.05.}
}
\description{
This function provides an easy interface for running the ACGCA model.
}
\examples{
ACGCArun()
}
\keyword{IBM}
