% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_error_func.r
\name{errorcode}
\alias{errorcode}
\title{ACGCA error function}
\usage{
errorcode(binerror, nsets)
}
\arguments{
\item{binerror}{a vector of errors in integer form}

\item{nsets}{the number of parameter sets processed}
}
\description{
This function reports error codes.
}
