% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACGCA_call_met.R
\name{growthloopR}
\alias{growthloopR}
\title{ACGCA model function}
\usage{
growthloopR(sparms2, gparms2, r0)
}
\arguments{
\item{r0}{The starting radius. Defaults to 0.05.}

\item{sparms}{A vector containing the parameters for the simulation.
\enumerate{
   \hmax{tt}
   \phih{dd}
 }}

\item{gparms}{A vector containing values that control simulation behavior.}
}
\description{
This function provides an easy interface for running the ACGCA model.
}
\keyword{IBM}
