% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACGCA_call_met.R
\name{growthloopR}
\alias{growthloopR}
\title{ACGCA model function}
\usage{
growthloopR(sparms, r0 = 0.05, parmax = 2060, years = 50, steps = 16,
  breast.height = 1.37, tolerance = 1e-05, fulloutput = FALSE)
}
\arguments{
\item{sparms}{A vector containing the parameters for the simulation.
\describe{
   \item{hmax}{Maximum tree height}
   \item{phih}{Slope of H vs. r curve at r = 0 m}
   \item{eta}{Relative height at which trunk transitions from paraboloid to
    cone}
   \item{swmax}{Maximum sapwood width}
   \item{lamdas}{Proportionality between BT and BO for sapwood}
   \item{lamdah}{Proportionality between BT and BO for heartwood}
   \item{rho}{Wood density}
   \item{rhomin}{Not in use. Dissabled in C code}
   \item{f2}{Leaf area-to-xylem conducting area ratio}
   \item{f1}{Fine root area-to-leaf area ratio}
   \item{gamma}{Maximum storage capacity of living sapwood cells}
   \item{gammaw}{(Inverse) density of sapwood structural tissue}
   \item{gammax}{Xylem conducting area-to-sapwood area ratio}
   \item{cgl}{Construction costs of producing leaves}
   \item{cgr}{Construction costs of producing fine roots}
   \item{cgw}{Construction costs of producing sapwood}
   \item{deltal}{Labile carbon storage capacity of leaves}
   \item{deltar}{Labile carbon storage capacity of fine roots}
   \item{sl}{Senescence rate of leaves}
   \item{sla}{Specific leaf area}
   \item{sr}{Senescence rate of fine roots}
   \item{so}{Senescence rate of coarse roots and branches}
   \item{rr}{Average fine root radius}
   \item{rhor}{Tissue density of fine roots}
   \item{rml}{Maintenance respiration rate of leaves}
   \item{rms}{Maintenance respiration rate of sapwood}
   \item{rmr}{Maintenance respiration rate of fine roots}
   \item{etaB}{Relative height at which trunk transitions from a neiloid to
    paraboloid}
   \item{k}{Crown light extinction coefficient}
   \item{epsg}{Radiation-use-efficiency}
   \item{m}{Maximum relative crown depth}
   \item{alpha}{Crown Curvature parameter}
   \item{R0}{Maximum potential crown radius of a tree with diameter at
    breast height of 0m (i.e., for a tree that is exactly 1.37 m tall}
   \item{R40}{Maximum potential crown radius of a tree with diameter at
    brrest height or 0.4m (40 cm)}
   \item{drinit}{NA}
   \item{drcrit}{NA}
 }}

\item{r0}{The starting radius. Defaults to 0.05}

\item{parmax}{The maximum yearly iradiance, defaults to 2060}

\item{years}{The number of years to run the simulation, defaults to 50.}

\item{steps}{The number of time steps per year, defaults to 16.}

\item{breast.height}{The height DBH is taken at, defaults to 1.37.}

\item{tolerance}{The tolerance for the algorithm that balances excess labile
carbon in the difference equations describing carbon dynamics of a healthy
tree (Ogle and Pacala, 2009). The default is 0.00001 and should likely not
be changed.}

\item{fulloutput}{Is the full output desired if so set this to TRUE. The
default is FALSE.}
}
\value{
Function output:
\describe{
   \item{p}{sparms, input parameters}
   \item{gp}{Vector: (timestep, years, tolerance, breast.height,parmax)}
   \item{r0}{The starting radius.}
   \item{t}{}
}

p

gp

r0

t

h

hh

hC

hB

hBH

r

rB

rC

rBH

sw

vts

vt

vth

sa

la

ra

dr

xa

bl

br

bt

bts

bth

boh

bos

bo

bs

cs

clr

fl

fr

ft

fo

rfl

rfr

rfs

egrow

ex

rtrans

light

nut

deltas

LAI

status

lenvars

errorind

growth_st
}
\description{
This function provides an easy interface for running the ACGCA model.
}
\keyword{IBM}
